+++ HISTORY.md
@@ -1,16 +1,19 @@
 # History
 
+## 0.0.4 (02-28-2024)
+
+  - Removed external data fetch for list of unsupported features in prep for adding display of features that are semi-supported.
+  - Added command line help **python merakicat.py help**
+  - Still not yet released on PyPI.
+
 ## 0.0.3 (02-27-2024)
 
   - Created command line option vs BOT.
   - Just try **python merakicat.py convert host <host or ip address> to <meraki network>**
-  - Not yet released on PyPI.
 
 ## 0.0.2 (02-26-2024)
 
-  - Updated some messages.
   - Not yet released on PyPI.
 
 ## 0.0.1 (02-26-2024)
 
-  - Not yet released on PyPI.